add openfga checks

+++ src/documents/document.router.ts
@@ -1,6 +1,12 @@
 import express from "express";
 import { validateAccessToken } from "../middleware/auth0.middleware";
-import { checkRequiredPermissions } from "../middleware/openfga.middleware";
+import {
+  checkRequiredPermissions,
+  createPermission,
+  deletePermission,
+  getPermission,
+  updatePermission,
+} from "../middleware/openfga.middleware";
 import {
   deleteDocumentById,
   findDocumentById,
@@ -8,26 +14,23 @@ import {
   saveDocument,
   updateDocument,
 } from "./document.service";
+import { create } from "ts-node";
 
 export const documentRouter = express.Router();
 
-documentRouter.get(
-  "/",
-  //validateAccessToken,
-  async (req, res, next) => {
-    try {
-      const documents = await getAllDocuments();
-      res.status(200).json(documents);
-    } catch (error) {
-      next(error);
-    }
+documentRouter.get("/", validateAccessToken, async (req, res, next) => {
+  try {
+    const documents = await getAllDocuments();
+    res.status(200).json(documents);
+  } catch (error) {
+    next(error);
   }
-);
+});
 
 documentRouter.post(
   "/",
-  //validateAccessToken,
-  //checkRequiredPermissions(["member", "team:admin"]),
+  validateAccessToken,
+  checkRequiredPermissions(createPermission),
   async (req, res, next) => {
     try {
       const document = await saveDocument(req.body);
@@ -40,8 +43,8 @@ documentRouter.post(
 
 documentRouter.put(
   "/:id",
-  //validateAccessToken,
-  //checkRequiredPermissions(["member", "team:admin"]),
+  validateAccessToken,
+  checkRequiredPermissions(updatePermission),
   async (req, res, next) => {
     try {
       const document = await updateDocument(req.params.id, req.body);
@@ -58,8 +61,8 @@ documentRouter.put(
 
 documentRouter.delete(
   "/:id",
-  //validateAccessToken,
-  //checkRequiredPermissions(["member", "team:admin"]),
+  validateAccessToken,
+  checkRequiredPermissions(deletePermission),
   async (req, res, next) => {
     try {
       const document = await deleteDocumentById(req.params.id);
@@ -67,7 +70,7 @@ documentRouter.delete(
         res.status(404).json({ message: "Document not found" });
         return;
       }
-      res.status(200);
+      res.status(200).send();
     } catch (error) {
       next(error);
     }
@@ -76,7 +79,8 @@ documentRouter.delete(
 
 documentRouter.get(
   "/:id",
-  //validateAccessToken,
+  validateAccessToken,
+  checkRequiredPermissions(getPermission),
   async (req, res, next) => {
     try {
       const document = await findDocumentById(req.params.id);
